feat(worker): add optimization flags (#61)
* feat(worker): add optimization flags

This commit enables users to pass optimization flags to the AI inference
pipelines. Due to startup delays we currently only pass these flags to
containers that are started warm.

* fix: decrease I2V SFAST warmup 'decode_chunk_size'

This commit decreases the `decode_chunk_size` used in the warmup passes
when SFAST is enabled. This was done to prevent `CUDA out of memory` errors
from occuring.

* fix: fix SFAST container timeout

This commit ensures that a longer timeout is used in the Runner
Container context so that pipelines with the SFAST optimization enabled
have enough time to startup. It also renames the `StringBool` type to
the more descriptive `EnvValue`.