feat(runner): disable DEEPCACHE for lightning/turbo models (#93)
* feat(runner): disable DEEPCACHE for lightning/turbo models

This commit ensures that people can not use the deepcache optimization
with lightning and turbo models. As explained in
https://github.com/livepeer/ai-worker/issues/82#issuecomment-2141983903
this optimization does not offer any speedup for these models while it
does reduce image quality.

* fix: ensure correct lighthing/turbo matching

This commit ensures that Lightning/Turbo model names with both the `-`
and `_` are matched when checking whether the orchestrator loaded a
Lightning/Turbo model.