feat(pipeline): add upscale image support to AI pipelines (#96)
* add upscale image support using stabilityai/stable-diffusion-x4-upscaler model

* Add host port for upscale pipeline

* added model download to the dl_checkpoints.sh

* fix: fix some small bugs and improve formatting

This commit fixes some small bugs and improves the code formatting so
that it is more in line with the other pipelines.

* fix: ensure upscaling OpenAPI spec gets created

This commit ensures that the `gen_openapi.py` file also creates the
OpenAPI spec for the upscaling route. It also updates the Golang client
bindings.

* fix(worker): fix incorrect automatic client types

This commit ensures that the right client request and response types
are used.

---------

Co-authored-by: Elite Encoder <john@eliteencoder.net>
Co-authored-by: Rick Staa <rick.staa@outlook.com>