runner/docker: upgrade to cuda 12.3 on base image
The DepthAnything ComfyUI nodes require tensorrt v10.x which is not compatible
with cuda 12.2 (compatible with tensorrt v8.x) -- cuda v12.3 is compatible with
tensorrt 10.x so upgrade it here. The existing streamdiffusion and liveportrait
pipelines should still run but needs to be tested later once the ingress/egress
functionality is in place.