runner/live: Make sure to resize images outside of infer process
On the LiveKit PoC we resized the images on the agent code, and
were missing that part from this infer.py piece. This is probably
one reason for the inference to be lower on the ai-worker as the
resize still happened, but it happened on the same process as we
run the model so it was competing for the same Python GIL (and
resizing on CPU might be some relevant time).